backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
spacy==3.7.5
nltk==3.8.1
scikit-learn==1.4.2
textstat==0.7.3
numpy==1.26.4
pypdf==6.10.1
groq==0.9.0
python-dotenv==1.0.1
